ci: add e2e test and remove yml workflow test file

+++ app/Services/Auth/OtpService.php
@@ -122,7 +122,7 @@ class OtpService
             return 0;
         }
 
-        return max(0, (int) $cooldownTimestamp - now()->timestamp);
+        return max(0, ((int) $cooldownTimestamp) - ((int) now()->timestamp));
     }
 
     /**
